requirements.txt
Django==4.0
psycopg2-binary==2.9.9
django-filter==23.5
djangorestframework==3.14.0
python-dotenv==1.0.0
